# backend/dist/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/vk": {
      "post": {
        "operationId": "AuthController_vkAuth",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VkAuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешная авторизация"
          },
          "401": {
            "description": "Неверные данные"
          }
        },
        "summary": "VK авторизация",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный вход"
          },
          "401": {
            "description": "Неверные учетные данные"
          }
        },
        "summary": "Вход в систему",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Пользователь создан"
          },
          "400": {
            "description": "Неверные данные"
          }
        },
        "summary": "Регистрация пользователя",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Профиль пользователя"
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить профиль пользователя",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/businesses": {
      "post": {
        "operationId": "BusinessController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBusinessDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Бизнес создан"
          },
          "400": {
            "description": "Неверные данные"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Создать бизнес",
        "tags": [
          "Business"
        ]
      },
      "get": {
        "operationId": "BusinessController_findAllByOwner",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список бизнесов"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить все бизнесы владельца",
        "tags": [
          "Business"
        ]
      }
    },
    "/api/v1/businesses/{id}": {
      "get": {
        "operationId": "BusinessController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID бизнеса",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Бизнес найден"
          },
          "404": {
            "description": "Бизнес не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить бизнес по ID",
        "tags": [
          "Business"
        ]
      },
      "patch": {
        "operationId": "BusinessController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID бизнеса",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBusinessDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Бизнес обновлен"
          },
          "404": {
            "description": "Бизнес не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Обновить бизнес",
        "tags": [
          "Business"
        ]
      }
    },
...
    "/api/v1/businesses/{businessId}/bookings": {
      "post": {
        "operationId": "BookingsController_create",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Bookings"
        ]
      },
      "get": {
        "operationId": "BookingsController_findAll",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Дата начала периода (ISO строка)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Дата окончания периода (ISO строка)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "staffId",
            "required": false,
            "in": "query",
            "description": "ID сотрудника",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Статус записи",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "CONFIRMED",
                "CANCELLED",
                "COMPLETED",
                "NO_SHOW"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Bookings"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/bookings/available-slots": {
      "get": {
        "operationId": "BookingsController_getAvailableSlots",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "serviceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "staffId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Bookings"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/bookings/{id}": {
      "get": {
        "operationId": "BookingsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Bookings"
        ]
      },
      "patch": {
        "operationId": "BookingsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Bookings"
        ]
      },
      "delete": {
        "operationId": "BookingsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Bookings"
        ]
      }
    },
...
    "/api/v1/businesses/{businessId}/availability": {
      "get": {
        "operationId": "AvailabilityController_getAvailability",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "serviceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "staffId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Availability"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/availability/exceptions": {
      "get": {
        "operationId": "AvailabilityController_getExceptions",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Availability"
        ]
      },
      "post": {
        "operationId": "AvailabilityController_createException",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Availability"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/availability/exceptions/{exceptionId}": {
      "delete": {
        "operationId": "AvailabilityController_deleteException",
        "parameters": [
          {
            "name": "exceptionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Availability"
        ]
      }
    }
  },
  "info": {
    "title": "Mastro API",
    "description": "API для системы управления записями",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "VkAuthDto": {
        "type": "object",
        "properties": {
          "vkToken": {
            "type": "string",
            "description": "VK токен для авторизации"
          }
        },
        "required": [
          "vkToken"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email пользователя"
          },
          "password": {
            "type": "string",
            "description": "Пароль пользователя"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email пользователя"
          },
          "password": {
            "type": "string",
            "description": "Пароль пользователя"
          },
          "vkId": {
            "type": "string",
            "description": "VK ID пользователя"
          },
          "role": {
            "type": "string",
            "description": "Роль пользователя",
            "enum": [
              "OWNER",
              "STAFF"
            ]
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "CreateBusinessDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название бизнеса",
            "maxLength": 100
          },
          "address": {
            "type": "string",
            "description": "Адрес бизнеса",
            "maxLength": 500
          },
          "timezone": {
            "type": "string",
            "description": "Часовой пояс"
          },
          "phone": {
            "type": "string",
            "description": "Телефон"
          },
          "email": {
            "type": "string",
            "description": "Email"
          },
          "website": {
            "type": "string",
            "description": "Веб-сайт"
          },
          "description": {
            "type": "string",
            "description": "Описание бизнеса"
          },
          "workingHours": {
            "type": "object",
            "description": "Рабочие часы"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateBusinessDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название бизнеса"
          },
          "address": {
            "type": "string",
            "description": "Адрес бизнеса",
            "maxLength": 500
          },
          "timezone": {
            "type": "string",
            "description": "Часовой пояс"
          },
          "phone": {
            "type": "string",
            "description": "Телефон"
          },
          "email": {
            "type": "string",
            "description": "Email"
          },
          "website": {
            "type": "string",
            "description": "Веб-сайт"
          },
          "description": {
            "type": "string",
            "description": "Описание бизнеса"
          },
          "workingHours": {
            "type": "object",
            "description": "Рабочие часы"
          }
        }
      },
...
      "CreateBookingDto": {
        "type": "object",
        "properties": {
          "serviceId": {
            "type": "string",
            "description": "ID услуги"
          },
          "staffId": {
            "type": "string",
            "description": "ID сотрудника"
          },
          "startTs": {
            "type": "string",
            "description": "Время начала записи (ISO строка)"
          },
          "clientId": {
            "type": "string",
            "description": "ID клиента"
          },
          "client": {
            "type": "object",
            "description": "Данные клиента (если создается новый)"
          }
        },
        "required": [
          "serviceId",
          "staffId",
          "startTs"
        ]
      },
      "UpdateBookingDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Статус записи",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "CANCELLED",
              "COMPLETED",
              "NO_SHOW"
            ]
          }
        }
      },
...
    }
  }
}